requirements.txt
numpy==1.26.4
pillow==10.3.0
pip==24.0
